test(cloud): split the packaged dispatcher contract into its own file
The rebase onto current main lands on a package-electron-runtime-contract.test.mjs
that already sits at the max-lines ceiling (598 of 600 effective lines), so adding
the packaged-undici assertion there fails `pnpm lint`. Move that one contract into
its own file. The original suite returns to its main content unchanged, which also
removes the only conflict surface this branch had against main.
